+++ cloud/client.go
@@ -96,6 +96,7 @@ func (it *internalClient) does(method string, request *Request) *Response {
 		httpRequest.TransferEncoding = []string{request.TransferEncoding}
 	}
 	httpRequest.Header.Add("robocorp-installation-id", xviper.TrackingIdentity())
+	httpRequest.Header.Add("User-Agent", common.UserAgent())
 	for name, value := range request.Headers {
 		httpRequest.Header.Add(name, value)
 	}